backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
razorpay==1.4.2
boto3==1.34.0
sendgrid==6.11.0
python-dotenv==1.0.0
dj-database-url==2.1.0

